backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beanie==1.25.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.109.2
google-auth==2.23.4
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.33.1
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
gunicorn==21.2.0